web-ui/requirements.txt
Flask
Pillow
pyyaml

